Revert "change distances to floats for accuracy"
This reverts commit 772bb272576083f7afe6cbcde78ed6db1bed0503.

+++ src/main/java/DeliveryV1/RouteGroup.java
@@ -22,7 +22,7 @@ public class RouteGroup
         return totalDistance;
     }
 
-    public float CalculateTotalDistance(float[][] Distances, int[][] Coordinates)
+    public int CalculateTotalDistance(int[][] Distances, int[][] Coordinates)
     {
         totalDistance = 0;
         for (int i = 0; i < GroupSize; i++)
